DailymotionApp/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
blueprint==3.4.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
countdown==0.1.0
distlib==0.3.1
Django==3.1.3
filelock==3.0.12
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.1
Werkzeug==1.0.1
